feat(website): give license pricing its own section on /pricing
Moves the tier table out of the resource-costs card into a standalone
"License pricing" section with its own description, and reverts the
slot added to the shared platform pricing section.
