+++ apps/website/src/pages/pricing.astro
@@ -5,7 +5,7 @@ import WhatsIncludedSection from '../components/pricing/WhatsIncludedSection.vue
 import PricingFaq from '../components/pricing/PricingFaq.astro'
 import ResourceCostsSection from '../templates/platform/PricingSection.vue'
 import LookingForElseSection from '../components/pricing/LookingForElseSection.vue'
-import MinimaxLicenseCostBand from '../components/pricing/MinimaxLicenseCostBand.vue'
+import MinimaxLicensePricingSection from '../components/pricing/MinimaxLicensePricingSection.vue'
 import MinimaxLicenseHeroSection from '../components/common/MinimaxLicenseHeroSection.vue'
 import { pricingOffers } from '../config/pricing'
 import { t } from '../i18n/translations'
@@ -50,9 +50,8 @@ const productId = jsonLdId(url, 'product')
     locale={locale}
     heading={t('pricing.resourceCosts.heading', locale)}
     note={t('pricing.resourceCosts.note', locale)}
-  >
-    <MinimaxLicenseCostBand locale={locale} />
-  </ResourceCostsSection>
+  />
+  <MinimaxLicensePricingSection locale={locale} />
   <LookingForElseSection locale={locale} />
   <PricingFaq locale={locale} />
 </BaseLayout>
