+++ apps/website/src/pages/zh-CN/pricing.astro
@@ -5,7 +5,7 @@ import WhatsIncludedSection from '../../components/pricing/WhatsIncludedSection.
 import PricingFaq from '../../components/pricing/PricingFaq.astro'
 import ResourceCostsSection from '../../templates/platform/PricingSection.vue'
 import LookingForElseSection from '../../components/pricing/LookingForElseSection.vue'
-import MinimaxLicenseCostBand from '../../components/pricing/MinimaxLicenseCostBand.vue'
+import MinimaxLicensePricingSection from '../../components/pricing/MinimaxLicensePricingSection.vue'
 import MinimaxLicenseHeroSection from '../../components/common/MinimaxLicenseHeroSection.vue'
 import { pricingOffers } from '../../config/pricing'
 import { t } from '../../i18n/translations'
@@ -53,9 +53,8 @@ const productId = jsonLdId(url, 'product')
     locale="zh-CN"
     heading={t('pricing.resourceCosts.heading', 'zh-CN')}
     note={t('pricing.resourceCosts.note', 'zh-CN')}
-  >
-    <MinimaxLicenseCostBand locale="zh-CN" />
-  </ResourceCostsSection>
+  />
+  <MinimaxLicensePricingSection locale="zh-CN" />
   <LookingForElseSection locale="zh-CN" />
   <PricingFaq locale="zh-CN" />
 </BaseLayout>
